Full testing environment
Former-commit-id: a87f2cc71b54a41d4fcea2eb45e5b89c98d42c52

+++ src/util/GameConnection.py
@@ -5,12 +5,14 @@ import threading
 import time
 from subprocess import Popen as new
 
+import subprocess
+
 import keyboard
 from websocket_server import WebsocketServer
 
 
 class GameConnection(threading.Thread):
-    def __init__(self, host = 'localhost', port = 9000):
+    def __init__(self, host = 'localhost', port = 9001):
         super().__init__()
 
         self.condition_object = threading.Condition()
@@ -23,7 +25,8 @@ class GameConnection(threading.Thread):
         self.response = None
 
         self.client = None
-        self.game_process = None
+        self.game_process: subprocess.Popen[str] = None
+        self.ai_process: subprocess.Popen[str] = None
 
     def run(self):
         self.server.run_forever()
@@ -48,15 +51,49 @@ class GameConnection(threading.Thread):
 
     def game_window_closed(self, client, server):
         print(f"Game Window Closed")
-        self.game_process.terminate()
-        self.game_process.wait()
+        if self.game_process:
+            self.game_process.terminate()
+            self.game_process.wait()
+
+        if self.ai_process:
+            self.ai_process.terminate()
+            self.ai_process.wait()
+
         self.client = None
 
     def start_game(self, game_path = '../science_birds/win-new/ScienceBirds.exe'):
         self.game_process = new(game_path, shell = False)
 
+    def startAi(self, ai_process = '../ai/Naive-Agent-standalone-Streamlined.jar'):
+
+        if self.ai_process:
+            self.ai_process.terminate()
+            self.ai_process.wait()
+
+        message = [0, 'aimodus', 'true']
+        self.send(message)
+        self.wait_for_response()
+        self.ai_process = new(ai_process)
+
+    def stopAI(self):
+
+        message = [0, 'aimodus', 'true']
+        self.send(message)
+        self.wait_for_response()
+
+        if self.ai_process:
+            self.ai_process.terminate()
+            self.ai_process.wait()
+
+    def stop(self):
+        self.server.shutdown_gracefully()
+        if self.game_process:
+            self.game_process.terminate()
+        if self.ai_process:
+            self.ai_process.terminate()
+
     def new_client(self, client, server: WebsocketServer):
-        print("New game window connected")
+        print(f"New game window connected {client}")
         self.client = client
         self.condition_object.acquire()
         self.condition_object.notify()
@@ -89,17 +126,23 @@ class GameConnection(threading.Thread):
             self.wait_for_response()
             time.sleep(0.2)
 
+    def wait_till_all_level_played(self):
+        while True:
+            if self.response is not None and 'data' in self.response[1] and self.response[1]['data'] == "True":
+                break
+
+            print("Wait for level played")
+            message = [0, 'alllevelsplayed']
+            self.send(message)
+            self.wait_for_response()
+            time.sleep(1)
+
     def change_level(self, index = 0):
         message = [0, 'selectlevel', {'levelIndex': index}]
         self.send(message)
         self.wait_for_response()
 
-    def stop(self):
-        self.server.shutdown_gracefully()
-        if self.game_process:
-            self.game_process.terminate()
-
-    def getData(self):
+    def get_data(self):
         message = [0, 'getdata']
         self.send(message)
         self.wait_for_response()
@@ -108,23 +151,25 @@ class GameConnection(threading.Thread):
 
 if __name__ == '__main__':
     game_connection = GameConnection()
-    game_connection.start()
 
-    print("Start game")
-    game_connection.start_game()
+    try:
+        game_connection.start()
 
-    print("Start wait")
-    game_connection.wait_for_game_window()
-    # time.sleep(2.4)
+        # print("Start game")
+        # game_connection.start_game()
+        game_connection.wait_for_game_window()
 
-    print("Change Level")
-    game_connection.change_level(index = 3)
+        print("Change Level")
+        game_connection.change_level(index = 3)
 
-    time.sleep(5)
+        print("Start AI")
+        game_connection.startAi()
 
-    print("Change Level")
-    game_connection.getData()
+        print("Change Level")
+        game_connection.get_data()
 
-    print("Wait for keyboard: p")
-    keyboard.wait("p")
-    game_connection.stop()
+        print("Wait for keyboard: p")
+        keyboard.wait("p")
+        game_connection.stop()
+    except KeyboardInterrupt:
+        game_connection.stop()
